Make compatible with graphql-batch
When a field resolver method returns the result of a batch loader, it
returns a promise. `graphl-cache` was attempting to write promises,
or more accurately promise-like, lazily resolved values that `graphl-batch`
[tells](https://github.com/Shopify/graphql-batch/blob/75bfef9b98faa15b59ee5df7f6ba1253083d7ee2/lib/graphql/batch.rb#L40)
`graphql-ruby` to wrap the promises it returns from batch loaders in,
to the cache instead of waiting for a cacheable value to be ready.
Trying to cache the lazy value would result in nothing being written to cache,
and an error being logged.

This makes sure that writing to cache waits until a cacheable value is
ready.

Fixes #55